Add files via upload
This is Music Player created using HTML, CSS and Javascript. On clicking the next button, the next song begins to play and on clicking the previous button, the previous song begins to play. On pressing the next button after the third song , which is the last song, the first song plays again and vice-versa. This can help us to shortlist our favourite songs and club them together.